bubble_sort

Function bubble_sort

pub fn bubble_sort<T: PartialOrd>(lista: &mut [T])

ordena un slice utilizando el algoritmo de burbuja (Bubble Sort)

Este método es estable y tiene una optimización de “salida temprana” si la lista se ordena antes de completar todas las interacciones.

Parámetros

lista - Una referencia mutable a un slice de elementos que implementan PartialOrd.

Ejemplo

let mut numeros = vec![64, 34, 25, 12, 22, 11, 90];
bubble_sort(&mut numeros);
assert_eq!(numeros, vec![11, 12, 22, 25, 34, 64, 90]);

Complejidad

Tiempo: O(n^2) en el peor caso, O(n) en el mejor caso (cuando la lista ya está ordenada).

Espacio: O(1)